+++ src/bitset_interface.rs
@@ -6,18 +6,8 @@ use crate::ops::BitSetOp;
 use crate::bit_block::BitBlock;
 use crate::cache::ReduceCache;
 use crate::config::{DefaultBlockIterator, Config, DefaultIndexIterator};
-use crate::bitset_app::BitSetApp;
+use crate::apply::Apply;
 use crate::reduce::Reduce;
-
-/// BitSet that have only non-empty elements in hierarchy blocks.
-/// (There could be no pointers to empty blocks in hierarchy)
-/// 
-/// This is usually by default, but some virtual bitsets may break this rule,
-/// for example all union operations are not `TrustedHierarchy`.
-pub trait TrustedHierarchy {
-    /// O(1)
-    fn is_empty(&self) -> bool;
-}
 
 // We have this separate trait with Config, to avoid making LevelMasks public.
 pub trait BitSetBase {
@@ -118,7 +108,6 @@ impl<'a, T: LevelMasks> LevelMasks for &'a T {
         <T as LevelMasks>::data_mask(self, level0_index, level1_index)
     }
 }
-
 impl<'a, T: LevelMasksExt> LevelMasksExt for &'a T {
     type Level1Blocks = T::Level1Blocks;
 
@@ -210,11 +199,23 @@ pub trait BitSetInterface
     fn contains(&self, index: usize) -> bool;
 }
 
+/// BitSet that have no pointers to empty blocks in hierarchy.
+/// 
+/// This is usually by default, but some virtual bitsets may break this rule,
+/// for example all union operations are not `TrustedHierarchy`.
+pub trait TrustedHierarchy: BitSetInterface {
+    /// O(1)
+    #[inline]
+    fn is_empty(&self) -> bool {
+        self.level0_mask().is_zero()
+    }
+}
+
 macro_rules! impl_all {
     ($macro_name: ident) => {
         $macro_name!(impl<Conf> for BitSet<Conf> where Conf: Config);
         $macro_name!(
-            impl<Op, S1, S2> for BitSetApp<Op, S1, S2>
+            impl<Op, S1, S2> for Apply<Op, S1, S2>
             where
                 Op: BitSetOp,
                 S1: LevelMasksExt<Conf = S2::Conf>,
@@ -235,7 +236,7 @@ macro_rules! impl_all_ref {
     ($macro_name: ident) => {
         $macro_name!(impl<'a, Conf> for &'a BitSet<Conf> where Conf: Config);
         $macro_name!(
-            impl<'a, Op, S1, S2> for &'a BitSetApp<Op, S1, S2>
+            impl<'a, Op, S1, S2> for &'a Apply<Op, S1, S2>
             where
                 Op: BitSetOp,
                 S1: LevelMasksExt<Conf = S2::Conf>,